+++ src/library.f90
@@ -4,7 +4,7 @@
 ! -----------------------------------------------
 subroutine fb1(x,ww,w,res)
   implicit none
-  real (kind = 8) :: x, ww, w, res
+  real (kind = dp) :: x, ww, w, res
   ! ww == W;
   if (abs(x)<=ww) then 
 	res = 1.0d0
@@ -17,7 +17,7 @@ end subroutine fb1
 ! -----------------------------------------------
 subroutine fb2(y,ww,w,res)
   implicit none
-  real (kind = 8) :: y, ww, w, res
+  real (kind = dp) :: y, ww, w, res
   ! ww == W;  
   if (y<0.0d0) then
 	write(*,*) 'z coordinates should be positive for B3'
@@ -36,7 +36,7 @@ end subroutine fb2
 !------------------------------------------------------
 subroutine fb3(y,ww,w,res)
   implicit none
-  real (kind = 8) :: y, ww, w, res
+  real (kind = dp) :: y, ww, w, res
   ! ww == W;  
   if (y<0.0d0) then
 	write(*,*) 'z coordinates should be positive for B3'
@@ -61,13 +61,13 @@ subroutine vlm(xl,volume)
   ! B.D. 8/21/05
   !
   integer(kind=4) :: i
-  real(kind=8) :: volume
-  real(kind=8),dimension(nesd,nen) :: xl
+  real(kind = dp) :: volume
+  real(kind = dp),dimension(nesd,nen) :: xl
   integer(kind=4),dimension(8,8) :: it = reshape((/ &
     1,2,3,4,5,6,7,8, 2,3,4,1,6,7,8,5, 3,4,1,2,7,8,5,6, &
     4,1,2,3,8,5,6,7, 5,8,7,6,1,4,3,2, 6,5,8,7,2,1,4,3, &
     7,6,5,8,3,2,1,4, 8,7,6,5,4,3,2,1/),(/8,8/))
-  real(kind=8),dimension(8) :: bb
+  real(kind = dp),dimension(8) :: bb
   !
   do i=1,8
     bb(i) = xl(2,it(2,i))*(xl(3,it(6,i))-xl(3,it(3,i))+xl(3,it(5,i)) &
